+++ src/main/scala/org/podval/tools/backend/Logging.scala
@@ -1,6 +1,8 @@
 package org.podval.tools.backend
 
 import org.gradle.api.{GradleException, Project}
+import org.podval.tools.build.ScalaVersion
+import org.podval.tools.gradle.ScalaExtension
 import org.podval.tools.platform.IntelliJIdea
 
 abstract class Logging(project: Project) extends WithProject(project):
@@ -15,3 +17,10 @@ abstract class Logging(project: Project) extends WithProject(project):
 
   private def pluginMessage(message: String): String =
     s"Plugin 'org.podval.tools.scalajs' in $project: $message."
+
+  final protected def getScalaVersionFromScalaExtension: ScalaVersion = ScalaExtension
+    .findScalaVersion(project)
+    .getOrElse(error(
+      s"""Scala version data is not supported when Scala version is inferred from the Scala library dependency;
+         |set Scala version on the Scala plugin's extension instead: `scala.scalaVersion=...`""".stripMargin
+    ))