+++ src/main/scala/org/podval/tools/backend/SingleBackendProject.scala
@@ -3,7 +3,7 @@ package org.podval.tools.backend
 import org.gradle.api.Project
 import org.gradle.api.plugins.jvm.internal.JvmPluginServices
 import org.podval.tools.build.ScalaBackend
-import org.podval.tools.gradle.Sources
+import org.podval.tools.gradle.{Projects, Sources}
 import org.podval.tools.platform.IntelliJIdea
 
 final class SingleBackendProject(
@@ -26,14 +26,17 @@ final class SingleBackendProject(
       isRunningInIntelliJ
     )
 
-    setScalaVersionFromParentAndAddVersionSpecificSources()
+    // Set Scala version from parent.
+    Projects.parent(project).foreach(setScalaVersionFromParent)
+    
     backend.apply(project, jvmPluginServices, isRunningInIntelliJ)
     backend.registerTasks(project)
 
     project.afterEvaluate: (_: Project) =>
+      addVersionSpecificSources()
       sharedProjects.map(_.project).foreach(Sources.addShared(_, project, isRunningInIntelliJ))
       backend.afterEvaluate(
         project,
         projectScalaLibrary = extension.getScalaLibrary,
-        pluginScalaLibrary = extension.getPluginScalaLibrary
+        pluginScalaLibrary  = extension.getPluginScalaLibrary
       )